+++ src/block-kit/index.ts
@@ -8,10 +8,11 @@ export { Context } from './Context'
 export { Divider } from './Divider'
 export { File } from './File'
 export { Image } from './Image'
+export { Input } from './Input'
 export { Section, Field } from './Section'
 
-// Interactive components
-export { Button } from './interactive/Button'
+// Block elements
+export { Button } from './elements/Button'
 export {
   Select,
   SelectFragment,
@@ -21,9 +22,12 @@ export {
   UsersSelect,
   ConversationsSelect,
   ChannelsSelect,
-} from './interactive/Select'
-export { Overflow, OverflowItem } from './interactive/Overflow'
-export { DatePicker } from './interactive/DatePicker'
+} from './elements/Select'
+export { Overflow, OverflowItem } from './elements/Overflow'
+export { DatePicker } from './elements/DatePicker'
+
+// PlainTextInput won't provide because Input block has an usage as component.
+// export { PlainTextInput } from './elements/PlainTextInput'
 
 // Composition objects
 export { Confirm } from './composition/Confirm'